+++ packages/pyroscope-flamegraph/src/Toolbar.tsx
@@ -1,8 +1,17 @@
-import React, { ReactNode, isValidElement } from 'react';
+import React, {
+  ReactNode,
+  RefObject,
+  useState,
+  useRef,
+  useLayoutEffect,
+  isValidElement,
+  memo,
+} from 'react';
 import classNames from 'classnames/bind';
 import { faUndo } from '@fortawesome/free-solid-svg-icons/faUndo';
 import { faCompressAlt } from '@fortawesome/free-solid-svg-icons/faCompressAlt';
 import { faProjectDiagram } from '@fortawesome/free-solid-svg-icons/faProjectDiagram';
+import { faEllipsisV } from '@fortawesome/free-solid-svg-icons/faEllipsisV';
 import { Maybe } from 'true-myth';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import useResizeObserver from '@react-hook/resize-observer';
@@ -10,13 +19,11 @@ import useResizeObserver from '@react-hook/resize-observer';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import Button from '@webapp/ui/Button';
 // eslint-disable-next-line import/no-extraneous-dependencies
-import Dropdown, { MenuItem } from '@webapp/ui/Dropdown';
 import { Tooltip } from '@pyroscope/webapp/javascript/ui/Tooltip';
-import { FitModes, HeadMode, TailMode } from './fitMode/fitMode';
+import { FitModes } from './fitMode/fitMode';
 import SharedQueryInput from './SharedQueryInput';
 import type { ViewTypes } from './FlameGraph/FlameGraphComponent/viewTypes';
 import type { FlamegraphRendererProps } from './FlameGraph/FlameGraphRenderer';
-import CheckIcon from './FlameGraph/FlameGraphComponent/CheckIcon';
 import {
   TableIcon,
   TablePlusFlamegraphIcon,
@@ -30,33 +37,75 @@ import styles from './Toolbar.module.scss';
 
 const cx = classNames.bind(styles);
 
-export const TOOLBAR_MODE_WIDTH_THRESHOLD = 900;
+const DIVIDER_WIDTH = 5;
+const QUERY_INPUT_WIDTH = 175;
+const LEFT_MARGIN = 2;
+const RIGHT_MARGIN = 2;
+const TOOLBAR_SQUARE_WIDTH = 40 + LEFT_MARGIN + RIGHT_MARGIN;
+const MORE_BUTTON_WIDTH = 16;
 
-export type ShowModeType = ReturnType<typeof useSizeMode>;
+const calculateCollapsedItems = (
+  clientWidth: number,
+  collapsedItemsNumber: number,
+  itemsW: number[]
+) => {
+  const availableToolbarItemsWidth =
+    collapsedItemsNumber === 0
+      ? clientWidth - QUERY_INPUT_WIDTH - 5
+      : clientWidth - QUERY_INPUT_WIDTH - MORE_BUTTON_WIDTH - 5;
 
-export const useSizeMode = (target: React.RefObject<HTMLDivElement>) => {
-  const [size, setSize] = React.useState<'large' | 'small'>('large');
-
-  const calcMode = (width: number) => {
-    if (width < TOOLBAR_MODE_WIDTH_THRESHOLD) {
-      return 'small';
+  let collapsedItems = 0;
+  let visibleItemsWidth = 0;
+  itemsW.reverse().forEach((v) => {
+    visibleItemsWidth += v;
+    if (availableToolbarItemsWidth <= visibleItemsWidth) {
+      collapsedItems += 1;
     }
-    return 'large';
-  };
-
-  React.useLayoutEffect(() => {
-    if (target.current) {
-      const { width } = target.current.getBoundingClientRect();
-
-      setSize(calcMode(width));
-    }
-  }, [target.current]);
-
-  useResizeObserver(target, (entry: ResizeObserverEntry) => {
-    setSize(calcMode(entry.contentRect.width));
   });
 
-  return size;
+  return collapsedItems;
+};
+
+const useMoreButton = (
+  target: RefObject<HTMLDivElement>,
+  toolbarItemsWidth: number[]
+) => {
+  const [isCollapsed, setCollapsedStatus] = useState(true);
+  const [collapsedItemsNumber, setCollapsedItemsNumber] = useState(0);
+
+  useLayoutEffect(() => {
+    if (target.current) {
+      const { width } = target.current.getBoundingClientRect();
+      const collapsedItems = calculateCollapsedItems(
+        width,
+        collapsedItemsNumber,
+        toolbarItemsWidth
+      );
+      setCollapsedItemsNumber(collapsedItems);
+    }
+  }, [target.current, toolbarItemsWidth]);
+
+  const handleMoreClick = () => {
+    setCollapsedStatus((v) => !v);
+  };
+
+  useResizeObserver(target, (entry: ResizeObserverEntry) => {
+    const { width } = entry.target.getBoundingClientRect();
+    const collapsedItems = calculateCollapsedItems(
+      width,
+      collapsedItemsNumber,
+      toolbarItemsWidth
+    );
+
+    setCollapsedItemsNumber(collapsedItems);
+    setCollapsedStatus(true);
+  });
+
+  return {
+    isCollapsed,
+    handleMoreClick,
+    collapsedItemsNumber,
+  };
 };
 
 export interface ProfileHeaderProps {
@@ -85,7 +134,12 @@ export interface ProfileHeaderProps {
 
 const Divider = () => <div className={styles.divider} />;
 
-const Toolbar = React.memo(
+type ToolbarItemType = {
+  width: number;
+  el: ReactNode;
+};
+
+const Toolbar = memo(
   ({
     view,
     handleSearchChange,
@@ -102,47 +156,137 @@ const Toolbar = React.memo(
     sharedQuery,
     ExportData,
   }: ProfileHeaderProps) => {
-    const toolbarRef = React.useRef<HTMLDivElement>(null);
-    const showMode = useSizeMode(toolbarRef);
+    const toolbarRef = useRef<HTMLDivElement>(null);
 
-    return (
-      <div role="toolbar" ref={toolbarRef} data-mode={showMode}>
-        <div className={styles.navbar}>
-          <SharedQueryInput
-            showMode={showMode}
-            onHighlightChange={handleSearchChange}
-            highlightQuery={highlightQuery}
-            sharedQuery={sharedQuery}
-          />
-          <div className={styles['space-filler']} />
-          <FitMode
-            showMode={showMode}
-            fitMode={fitMode}
-            updateFitMode={updateFitMode}
-          />
+    const fitModeItem = {
+      el: (
+        <>
+          <FitMode fitMode={fitMode} updateFitMode={updateFitMode} />
           <Divider />
-          <ResetView isFlamegraphDirty={isFlamegraphDirty} reset={reset} />
+        </>
+      ),
+      width: TOOLBAR_SQUARE_WIDTH * 2 + DIVIDER_WIDTH,
+    };
+    const resetItem = {
+      el: <ResetView isFlamegraphDirty={isFlamegraphDirty} reset={reset} />,
+      width: TOOLBAR_SQUARE_WIDTH,
+    };
+    const focusOnSubtree = {
+      el: (
+        <>
           <FocusOnSubtree
             selectedNode={selectedNode}
             onFocusOnSubtree={onFocusOnSubtree}
           />
-          {enableChangingDisplay ? (
-            <>
-              <Divider />
-              <ViewSection
-                flamegraphType={flamegraphType}
-                showMode={showMode}
-                view={view}
-                updateView={updateView}
-              />
-            </>
-          ) : null}
-          {isValidElement(ExportData) ? (
+          <Divider />
+        </>
+      ),
+      width: TOOLBAR_SQUARE_WIDTH + DIVIDER_WIDTH,
+    };
+
+    const viewSectionItem = enableChangingDisplay
+      ? {
+          el: (
+            <ViewSection
+              flamegraphType={flamegraphType}
+              view={view}
+              updateView={updateView}
+            />
+          ),
+          // sandwich view is hidden in diff view
+          width: TOOLBAR_SQUARE_WIDTH * (flamegraphType === 'single' ? 5 : 4), // 1px is to display divider
+        }
+      : null;
+    const exportDataItem = isValidElement(ExportData)
+      ? {
+          el: (
             <>
               <Divider />
               {ExportData}
             </>
-          ) : null}
+          ),
+          width: TOOLBAR_SQUARE_WIDTH + DIVIDER_WIDTH,
+        }
+      : null;
+
+    const filteredToolbarItems = [
+      fitModeItem,
+      resetItem,
+      focusOnSubtree,
+      viewSectionItem,
+      exportDataItem,
+    ].filter((v) => v !== null) as ToolbarItemType[];
+    const toolbarItemsWidth = filteredToolbarItems.reduce(
+      (acc, v) => [...acc, v.width],
+      [] as number[]
+    );
+
+    const { isCollapsed, collapsedItemsNumber, handleMoreClick } =
+      useMoreButton(toolbarRef, toolbarItemsWidth);
+
+    const toolbarFilteredItems = filteredToolbarItems.reduce(
+      (acc, v, i) => {
+        const isHiddenItem = i < collapsedItemsNumber;
+
+        if (isHiddenItem) {
+          acc.hidden.push(v);
+        } else {
+          acc.visible.push(v);
+        }
+
+        return acc;
+      },
+      { visible: [] as ToolbarItemType[], hidden: [] as ToolbarItemType[] }
+    );
+
+    return (
+      <div role="toolbar" ref={toolbarRef}>
+        <div className={styles.navbar}>
+          <div>
+            <SharedQueryInput
+              width={QUERY_INPUT_WIDTH}
+              onHighlightChange={handleSearchChange}
+              highlightQuery={highlightQuery}
+              sharedQuery={sharedQuery}
+            />
+          </div>
+          <div>
+            <div className={styles.itemsContainer}>
+              {toolbarFilteredItems.visible.map((v, i) => (
+                // eslint-disable-next-line react/no-array-index-key
+                <div key={i} className={styles.item} style={{ width: v.width }}>
+                  {v.el}
+                </div>
+              ))}
+              {collapsedItemsNumber !== 0 && (
+                <Tooltip placement="top" title="More">
+                  <button
+                    onClick={handleMoreClick}
+                    className={cx({
+                      [styles.moreButton]: true,
+                      [styles.active]: !isCollapsed,
+                    })}
+                  >
+                    <FontAwesomeIcon icon={faEllipsisV} />
+                  </button>
+                </Tooltip>
+              )}
+            </div>
+          </div>
+          {!isCollapsed && (
+            <div className={styles.navbarCollapsedItems}>
+              {toolbarFilteredItems.hidden.map((v, i) => (
+                <div
+                  // eslint-disable-next-line react/no-array-index-key
+                  key={i}
+                  className={styles.item}
+                  style={{ width: v.width }}
+                >
+                  {v.el}
+                </div>
+              ))}
+            </div>
+          )}
         </div>
       </div>
     );
@@ -188,7 +332,7 @@ function ResetView({
 }) {
   return (
     <Tooltip placement="top" title="Reset View">
-      <div>
+      <span>
         <Button
           id="reset"
           disabled={!isFlamegraphDirty}
@@ -198,7 +342,7 @@ function ResetView({
         >
           <FontAwesomeIcon icon={faUndo} />
         </Button>
-      </div>
+      </span>
     </Tooltip>
   );
 }
@@ -206,75 +350,15 @@ function ResetView({
 function FitMode({
   fitMode,
   updateFitMode,
-  showMode,
 }: {
-  showMode: ShowModeType;
   fitMode: ProfileHeaderProps['fitMode'];
   updateFitMode: ProfileHeaderProps['updateFitMode'];
 }) {
-  let texts = {
-    label: '',
-    [HeadMode]: '',
-    [TailMode]: '',
-  };
-  let menuButtonClassName = '';
-  switch (showMode) {
-    case 'small':
-      texts = {
-        label: 'Fit',
-        [HeadMode]: 'Head',
-        [TailMode]: 'Tail',
-      };
-      menuButtonClassName = styles.fitModeDropdownSmall;
-      break;
-    case 'large':
-      texts = {
-        label: 'Prefer to Fit',
-        [HeadMode]: 'Head first',
-        [TailMode]: 'Tail first',
-      };
-      menuButtonClassName = styles.fitModeDropdownLarge;
-      break;
-    default: {
-      throw new Error('Wrong mode');
-    }
-  }
-
-  const menuOptions = [HeadMode, TailMode] as FitModes[];
-  const menuItems = menuOptions.map((mode) => (
-    <MenuItem key={mode} value={mode}>
-      <div className={styles.dropdownMenuItem} data-testid={mode}>
-        {texts[mode]}
-        {fitMode === mode ? <CheckIcon /> : null}
-      </div>
-    </MenuItem>
-  ));
-
   const isSelected = (a: FitModes) => fitMode === a;
-
-  if (showMode === 'small') {
-    return (
-      <Tooltip placement="top" title="Fit Mode">
-        <div>
-          <Dropdown
-            label={texts.label}
-            ariaLabel="Fit Mode"
-            value={texts[fitMode]}
-            onItemClick={(event) =>
-              updateFitMode(event.value as typeof fitMode)
-            }
-            menuButtonClassName={menuButtonClassName}
-          >
-            {menuItems}
-          </Dropdown>
-        </div>
-      </Tooltip>
-    );
-  }
 
   return (
     <>
-      <Tooltip placement="top" title={texts['HEAD']}>
+      <Tooltip placement="top" title="Head first">
         <Button
           onClick={() => updateFitMode('HEAD')}
           className={cx({
@@ -285,7 +369,7 @@ function FitMode({
           <HeadFirstIcon />
         </Button>
       </Tooltip>
-      <Tooltip placement="top" title={texts['TAIL']}>
+      <Tooltip placement="top" title="Tail first">
         <Button
           onClick={() => updateFitMode('TAIL')}
           className={cx({
@@ -344,72 +428,32 @@ const getViewOptions = (
 function ViewSection({
   view,
   updateView,
-  showMode,
   flamegraphType,
 }: {
-  showMode: ShowModeType;
   updateView: ProfileHeaderProps['updateView'];
   view: ProfileHeaderProps['view'];
   flamegraphType: ProfileHeaderProps['flamegraphType'];
 }) {
   const options = getViewOptions(flamegraphType);
 
-  const dropdownMenuItems = options.map((mode) => (
-    <MenuItem key={mode.value} value={mode.value}>
-      <div className={styles.dropdownMenuItem} data-testid={mode.value}>
-        {mode.label}
-        {view === mode.value ? <CheckIcon /> : null}
-      </div>
-    </MenuItem>
-  ));
-
-  const ViewSelect = (
-    <Tooltip placement="top" title="View Mode">
-      <div>
-        <Dropdown
-          label="View Mode"
-          ariaLabel="View Mode"
-          value={options.find((i) => i.value === view)?.label}
-          onItemClick={(event) => updateView(event.value)}
-          align="end"
-          menuButtonClassName={styles.viewModeDropdownButton}
-        >
-          {dropdownMenuItems}
-        </Dropdown>
-      </div>
-    </Tooltip>
+  return (
+    <div className={styles.viewType}>
+      {options.map(({ label, value, Icon }) => (
+        <Tooltip key={value} placement="top" title={label}>
+          <Button
+            data-testid={value}
+            onClick={() => updateView(value)}
+            className={cx({
+              [styles.toggleViewButton]: true,
+              selected: view === value,
+            })}
+          >
+            <Icon />
+          </Button>
+        </Tooltip>
+      ))}
+    </div>
   );
-
-  const ViewButtons = options.map(({ label, value, Icon }) => (
-    <Tooltip key={value} placement="top" title={label}>
-      <Button
-        onClick={() => updateView(value)}
-        className={cx({
-          [styles.toggleViewButton]: true,
-          selected: view === value,
-        })}
-      >
-        <Icon />
-      </Button>
-    </Tooltip>
-  ));
-
-  const decideWhatToShow = () => {
-    switch (showMode) {
-      case 'small': {
-        return ViewSelect;
-      }
-      case 'large': {
-        return ViewButtons;
-      }
-
-      default: {
-        throw new Error(`Invalid option: '${showMode}'`);
-      }
-    }
-  };
-
-  return <div className={styles.viewType}>{decideWhatToShow()}</div>;
 }
 
 export default Toolbar;