+++ webapp/javascript/pages/TagExplorerView.tsx
@@ -13,7 +13,6 @@ import TimelineChartWrapper, {
 } from '@webapp/components/TimelineChart/TimelineChartWrapper';
 import { FlamegraphRenderer } from '@pyroscope/flamegraph/src';
 import Dropdown, { MenuItem } from '@webapp/ui/Dropdown';
-import LoadingSpinner from '@webapp/ui/LoadingSpinner';
 import TagsSelector from '@webapp/pages/tagExplorer/components/TagsSelector';
 import TableUI, { useTableSort, BodyRow } from '@webapp/ui/Table';
 import useColorMode from '@webapp/hooks/colorMode.hook';
@@ -38,6 +37,7 @@ import { queryToAppName } from '@webapp/models/query';
 import PageTitle from '@webapp/components/PageTitle';
 import ExploreTooltip from '@webapp/components/TimelineChart/ExploreTooltip';
 import { getFormatter } from '@pyroscope/flamegraph/src/format/format';
+import { LoadingOverlay } from '@webapp/ui/LoadingOverlay';
 import { calculateMean, calculateStdDeviation, calculateTotal } from './math';
 import { PAGES } from './constants';
 import {
@@ -46,7 +46,6 @@ import {
   getTableIntegerSpaceLengthByColumn,
   formatValue,
 } from './formatTableData';
-
 // eslint-disable-next-line css-modules/no-unused-class
 import styles from './TagExplorerView.module.scss';
 import { formatTitle } from './formatTitle';
@@ -180,7 +179,12 @@ function TagExplorerView() {
     }
   }, [query]);
 
-  const { groupByTag, groupByTagValue, type } = tagExplorerView;
+  const {
+    groupByTag,
+    groupByTagValue,
+    groupsLoadingType,
+    activeTagProfileLoadingType,
+  } = tagExplorerView;
 
   useEffect(() => {
     if (from && until && query && groupByTagValue) {
@@ -202,7 +206,7 @@ function TagExplorerView() {
     groupsData: TimelineGroupData[];
     activeTagProfile?: Profile;
   } => {
-    switch (tagExplorerView.type) {
+    switch (tagExplorerView.groupsLoadingType) {
       case 'loaded':
       case 'reloading':
         const groups = Object.entries(tagExplorerView.groups).reduce(
@@ -218,10 +222,15 @@ function TagExplorerView() {
           [] as TimelineGroupData[]
         );
 
-        if (groups.length > 0) {
+        if (
+          groups.length > 0 &&
+          (activeTagProfileLoadingType === 'loaded' ||
+            activeTagProfileLoadingType === 'reloading') &&
+          tagExplorerView?.activeTagProfile
+        ) {
           return {
             groupsData: groups,
-            activeTagProfile: tagExplorerView?.activeTagProfile,
+            activeTagProfile: tagExplorerView.activeTagProfile,
           };
         }
 
@@ -311,6 +320,11 @@ function TagExplorerView() {
       activeTagProfile.metadata.units
     );
 
+  const dataLoading =
+    groupsLoadingType === 'loading' ||
+    groupsLoadingType === 'reloading' ||
+    activeTagProfileLoadingType === 'loading';
+
   return (
     <>
       <PageTitle title={formatTitle('Tag Explorer View', query)} />
@@ -331,9 +345,7 @@ function TagExplorerView() {
             handleGroupByTagValueChange={handleGroupByTagValueChange}
           />
           <div id={TIMELINE_WRAPPER_ID} className={styles.timelineWrapper}>
-            {type === 'loading' ? (
-              <LoadingSpinner />
-            ) : (
+            <LoadingOverlay active={dataLoading}>
               <TimelineChartWrapper
                 selectionType="double"
                 mode="multiple"
@@ -360,7 +372,7 @@ function TagExplorerView() {
                   />
                 )}
               />
-            )}
+            </LoadingOverlay>
           </div>
         </Box>
         <CollapseBox
@@ -370,15 +382,12 @@ function TagExplorerView() {
         >
           <div className={styles.statisticsBox}>
             <div className={styles.pieChartWrapper}>
-              {groups?.length ? (
-                <TotalSamplesChart
-                  formatter={formatter}
-                  filteredGroupsData={groups}
-                  profile={activeTagProfile}
-                />
-              ) : (
-                <LoadingSpinner />
-              )}
+              <TotalSamplesChart
+                formatter={formatter}
+                filteredGroupsData={groups}
+                profile={activeTagProfile}
+                isLoading={dataLoading}
+              />
             </div>
             <Table
               appName={appName.unwrapOr('')}
@@ -387,7 +396,7 @@ function TagExplorerView() {
               groupByTagValue={groupByTagValue}
               groupsData={groups}
               handleGroupByTagValueChange={handleGroupByTagValueChange}
-              isLoading={type === 'loading'}
+              isLoading={dataLoading}
               activeTagProfile={activeTagProfile}
               formatter={formatter}
             />
@@ -395,9 +404,7 @@ function TagExplorerView() {
         </CollapseBox>
         <Box>
           <div className={styles.flamegraphWrapper}>
-            {type === 'loading' ? (
-              <LoadingSpinner />
-            ) : (
+            <LoadingOverlay active={dataLoading}>
               <FlamegraphRenderer
                 showCredit={false}
                 profile={activeTagProfile}
@@ -416,7 +423,7 @@ function TagExplorerView() {
                   )
                 }
               />
-            )}
+            </LoadingOverlay>
           </div>
         </Box>
       </div>
@@ -599,7 +606,7 @@ function Table({
   const table = {
     headRow,
     ...(isLoading
-      ? { type: 'not-filled' as const, value: <LoadingSpinner /> }
+      ? { type: 'not-filled' as const, value: <LoadingOverlay active /> }
       : { type: 'filled' as const, bodyRows }),
   };
 
@@ -669,6 +676,12 @@ function ExploreHeader({
   const handleGroupByValueClick = (e: ClickEvent) => {
     handleGroupByTagValueChange(e.value);
   };
+
+  useEffect(() => {
+    if (tagKeys.length && !selectedTag) {
+      handleGroupByTagChange(tagKeys[0]);
+    }
+  }, [tagKeys, selectedTag]);
 
   return (
     <div className={styles.header} data-testid="explore-header">