+++ webapp/javascript/ui/Menu.tsx
@@ -3,7 +3,7 @@ import {
   MenuItem,
   ControlledMenu,
   useMenuState,
-  ClickEvent,
 } from '@szhsin/react-menu';
+import type { ClickEvent } from '@szhsin/react-menu';
 
 export { ControlledMenu, Menu, MenuItem, useMenuState, ClickEvent };